Move workspace document tests.
